Update Record Linkage to use the new dibbs package (#1609)
* Update Record Linkage to use the new dibbs package

* Move functions to linkage.utils
