+++ containers/record-linkage/Dockerfile
@@ -1,10 +1,6 @@
-FROM python:3.10-slim
+FROM ghcr.io/cdcgov/phdi/dibbs
 
 WORKDIR /code
-
-RUN apt-get update && \
-    apt-get upgrade -y && \
-    apt-get install -y git build-essential
 
 COPY ./requirements.txt /code/requirements.txt
 RUN pip install -r requirements.txt
