+++ containers/record-linkage/app/linkage/link.py
@@ -13,7 +13,7 @@ from app.linkage.mpi import BaseMPIConnectorClient
 from app.linkage.mpi import DIBBsMPIConnectorClient
 from app.linkage.utils import compare_strings
 from app.linkage.utils import datetime_to_str
-from app.utils import extract_value_with_resource_path
+from app.linkage.utils import extract_value_with_resource_path
 from pydantic import Field
 
 LINKING_FIELDS_TO_FHIRPATHS = {
