+++ containers/record-linkage/app/linkage/mpi.py
@@ -9,8 +9,8 @@ from typing import Literal
 
 from app.linkage.core import BaseMPIConnectorClient
 from app.linkage.dal import DataAccessLayer
+from app.linkage.utils import extract_value_with_resource_path
 from app.linkage.utils import load_mpi_env_vars_os
-from app.utils import extract_value_with_resource_path
 from sqlalchemy import and_
 from sqlalchemy import Select
 from sqlalchemy import select
