+++ containers/record-linkage/app/main.py
@@ -8,10 +8,10 @@ from app.linkage.algorithms import DIBBS_ENHANCED
 from app.linkage.link import add_person_resource
 from app.linkage.link import link_record_against_mpi
 from app.linkage.mpi import DIBBsMPIConnectorClient
-from app.utils import BaseService
 from app.utils import get_settings
 from app.utils import read_json_from_assets
 from app.utils import run_migrations
+from dibbs.base_service import BaseService
 from fastapi import Body
 from fastapi import Response
 from fastapi import status
