+++ containers/record-linkage/dev-requirements.txt
@@ -1,0 +1,2 @@
+pytest
+httpx
