+++ containers/record-linkage/requirements.txt
@@ -1,8 +1,4 @@
-typing_extensions==4.8.0
-uvicorn
-pydantic==1.10.9
-fastapi>=0.109.1
-httpx
+../dibbs
 pathlib
 pyway
 psycopg2-binary==2.9.9
